filemode を別ファイルに切り出し
